Implemented PR comments
`split_video` now returns a dictionary with stats about each created
video. When the main video splitting function is called, it will print
out each video that was created and its length in seconds.

+++ src/split_videos.py
@@ -10,6 +10,7 @@ While the os.environ written below should do it for you, it is possible that it 
 import os
 import argparse
 import logging
+import shutil
 
 from utils import pprint
 from data_preprocessing import detect_black_frames
@@ -38,13 +39,14 @@ def _split_videos_in_directory(directory_path:str, output_dir:str) -> None:
     '''
     os.makedirs(output_dir, exist_ok=True)
 
-    try:
-        for file in sorted(os.listdir(directory_path)):
-            if file.upper().endswith('.MP4'):
-                video_path = os.path.join(directory_path, file)
-                _split_video(video_path, output_dir)
-    except:
-        logger.error('Could not split videos in directory %s, removing specified output dir', directory_path)
+    for file in sorted(os.listdir(directory_path)):
+        if file.upper().endswith('.MP4'):
+            video_path = os.path.join(directory_path, file)
+            # _split_video handles errors with video creation
+            _split_video(video_path, output_dir)
+
+    # In case no videos are created (everything failed)
+    if not os.listdir(output_dir):
         os.rmdir(output_dir)
 
 
@@ -52,9 +54,21 @@ def _split_video(vid_path, out_path):
     os.makedirs(out_path, exist_ok=True)
     try:
         black_f_id, fps = detect_black_frames(vid_path)
-        split_video(vid_path, black_f_id, out_path, fps)
+        stats           = split_video(vid_path, black_f_id, out_path, fps)
+
+        logger.info('Succesfully split a video. With the following stats:')
+        for key, value in stats.items():
+            logger.info('%s has length %s seconds', key, value)
+
     except:
-        os.rmdir(out_path)
+        logger.error("Wasn't able to split video: %s. Removing directory %s", vid_path, out_path)
+
+        # Directory is empty
+        if os.listdir(out_path):
+            os.rmdir(out_path)
+
+        # Do nothing if dir non-empty since other (successfully created)
+        # videos could already be there. We don't want to remove those
 
 
 def main(args):
